Add ST STM32F746 disco board to tflite tutorial script
Currently tutorial script 'micro_tflite.py' assumes all boards with target
STM32F746 are the same. As a consequence once that target is selected the
script automatically defaults to the Nucleo board. However, the STM32F746
is also used on Discovery Kit boards (aka disco) which are quite similar
but have some differences, so Nucleo config and final image don't work on
the disco boards.

That commit adds a way to select a different dev board and adds comments
accordingly, informing how to use the script with the disco boards.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>