Adjust generate_project to generate a project based upon a MLF archive
Currently the tvm.micro.project.generate_project method only creates a
project directory using a GraphExecutorFactory. That's not convenient
for TVMC because the Relay program is compiled on a previous stage, so
only the MLF archive artifact persists after the 'tvmc compile' command
terminates.

Since generate_project() will be used after the compile stage by TVMC,
it needs to be create based upon a MLF archive instead of a Graph
executor instance, hence that commits modifies the generate_project()
to also accept a MLF archive path to generate a project.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>