[TVMC] Add build command to micro context
This commit adds the build command to the new micro TVMC context. It
allows one to build a project (generating an image ready to be flash to
a micro target, e.g. zephyr.elf) given a project directory usually
generated on a stage before using the 'tvmc create-project'. It uses
the new Project API.

Gustavo Romero <gustavo.romero@linaro.org>